work in progress

+++ includes/minirt.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minirt.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tvasilev <tvasilev@student.42.fr>          +#+  +:+       +#+        */
+/*   By: tfregni <tfregni@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/23 16:43:46 by tfregni           #+#    #+#             */
-/*   Updated: 2023/09/04 17:59:26 by tvasilev         ###   ########.fr       */
+/*   Updated: 2023/09/07 14:52:00 by tfregni          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -163,7 +163,7 @@ int		intersect_element(t_scene *scene, t_ray ray, int *color, float *min_t);
 /* SHADE */
 float	light_coeff(t_scene *scene, t_hit_record *hit);
 float	diffuse_reflection(t_scene *scene, t_vector n, t_vector p);
-int		cast_shadow(t_scene *scene, t_ray ray);
+float		cast_shadow(t_scene *scene, t_ray ray);
 
 /* ERROR HANDLING*/
 t_err	ft_error(char *msg, char *arg, int err_code, t_scene *scene);
